Tests - Flask Migrations to Fast API
