+++ tests/covalent_dispatcher_tests/_service/app_test.py
@@ -24,7 +24,7 @@ import json
 from datetime import datetime
 
 import pytest
-from flask import Flask
+from fastapi.testclient import TestClient
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session
 
@@ -33,22 +33,20 @@ from covalent._data_store.datastore import DataStore
 from covalent._data_store.models import Lattice
 from covalent._results_manager.result import Result
 from covalent_dispatcher._db.dispatchdb import DispatchDB
-from covalent_dispatcher._service.app import bp as dispatcher
+from covalent_ui.app import fastapi_app as fast_app
 
 DISPATCH_ID = "f34671d1-48f2-41ce-89d9-9a8cb5c60e5d"
-
-flask_app = Flask(__name__)
-flask_app.register_blueprint(dispatcher)
 
 
 @pytest.fixture
 def app():
-    yield flask_app
+    yield fast_app
 
 
 @pytest.fixture
-def client(app):
-    return app.test_client()
+def client():
+    with TestClient(fast_app) as c:
+        yield c
 
 
 @pytest.fixture
@@ -75,12 +73,12 @@ class MockDataStore:
 def test_submit(mocker, app, client):
     mocker.patch("covalent_dispatcher.run_dispatcher", return_value=DISPATCH_ID)
     response = client.post("/api/submit", data=json.dumps({}))
-    assert json.loads(response.data) == DISPATCH_ID
+    assert response.json() == DISPATCH_ID
 
 
 def test_cancel(app, client):
     response = client.post("/api/cancel", data=DISPATCH_ID.encode("utf-8"))
-    assert json.loads(response.data) == f"Dispatch {DISPATCH_ID} cancelled."
+    assert response.json() == f"Dispatch {DISPATCH_ID} cancelled."
 
 
 def test_db_path(mocker, app, client):
@@ -91,7 +89,8 @@ def test_db_path(mocker, app, client):
 
     mocker.patch.object(DispatchDB, "__init__", __init__)
     response = client.get("/api/db-path")
-    assert json.loads(response.data) == dbpath
+    result = response.json().replace("\\", "").replace('"', "")
+    assert result == dbpath
 
 
 def test_get_result(mocker, app, client, test_db, tmp_path):
@@ -111,8 +110,9 @@ def test_get_result(mocker, app, client, test_db, tmp_path):
 
     mocker.patch("covalent_dispatcher._service.app.result_from", return_value={})
     mocker.patch("covalent_dispatcher._service.app.workflow_db", test_db)
+    print(test_db)
     response = client.get(f"/api/result/{DISPATCH_ID}")
-    result = json.loads(response.data)
+    result = response.json()
     assert result["id"] == DISPATCH_ID
     assert result["status"] == str(Result.COMPLETED)
 
@@ -133,7 +133,7 @@ def test_get_result_503(mocker, app, client, test_db, tmp_path):
         session.commit()
     mocker.patch("covalent_dispatcher._service.app.result_from", side_effect=FileNotFoundError())
     mocker.patch("covalent_dispatcher._service.app.workflow_db", test_db)
-    response = client.get(f"/api/result/{DISPATCH_ID}")
+    response = client.get(f"/api/result/{DISPATCH_ID}?wait={False}&status_only={False}")
     assert response.status_code == 503
